requirements.txt
asyncgui==0.6.3
asynckivy==0.6.4
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
docutils==0.21.2
filetype==1.2.0
idna==3.10
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kivymd==1.1.1
materialyoucolor==2.0.10
pillow==11.2.1
Pygments==2.19.2
pypiwin32==223
pywin32==310
qrcode==8.2
requests==2.32.4
urllib3==2.5.0